Refactor performer urls into component
