+++ ui/v2.5/src/components/Shared/ExternalLinksButton.tsx
@@ -3,6 +3,8 @@ import { ExternalLink } from "./ExternalLink";
 import TextUtils from "src/utils/text";
 import { Icon } from "./Icon";
 import { IconDefinition, faLink } from "@fortawesome/free-solid-svg-icons";
+import { useMemo } from "react";
+import { faInstagram, faTwitter } from "@fortawesome/free-brands-svg-icons";
 
 export const ExternalLinksButton: React.FC<{
   icon?: IconDefinition;
@@ -45,5 +47,39 @@ export const ExternalLinksButton: React.FC<{
         ))}
       </Dropdown.Menu>
     </Dropdown>
+  );
+};
+
+export const ExternalLinkButtons: React.FC<{ urls: string[] | undefined }> = ({
+  urls,
+}) => {
+  const urlSpecs = useMemo(() => {
+    if (!urls?.length) {
+      return [];
+    }
+
+    const twitter = urls.filter((u) =>
+      u.match(/https?:\/\/(?:www\.)?twitter.com\//)
+    );
+    const instagram = urls.filter((u) =>
+      u.match(/https?:\/\/(?:www\.)?instagram.com\//)
+    );
+    const others = urls.filter(
+      (u) => !twitter.includes(u) && !instagram.includes(u)
+    );
+
+    return [
+      { icon: faLink, className: "", urls: others },
+      { icon: faTwitter, className: "twitter", urls: twitter },
+      { icon: faInstagram, className: "instagram", urls: instagram },
+    ];
+  }, [urls]);
+
+  return (
+    <>
+      {urlSpecs.map((spec, i) => (
+        <ExternalLinksButton key={i} {...spec} />
+      ))}
+    </>
   );
 };
